# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cost-RAG 工程造价咨询智能RAG系统 API
  description: |
    Cost-RAG是基于成本优化的检索增强生成系统，专为工程造价咨询行业设计。
    提供文档处理、智能问答、成本估算和多项目对比等核心功能。

    ## 主要特性
    - 🔍 支持多格式文档处理和智能检索
    - 💰 14级分部分项层级成本计算
    - 📊 多项目对比分析和相似性评估
    - 🤖 基于RAG的智能问答系统
    - 🚀 高性能异步处理架构

    ## 认证方式
    使用JWT Bearer Token进行认证，在请求头中添加：
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## 错误处理
    所有错误响应都遵循统一格式：
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "错误描述",
        "details": []
      }
    }
    ```
  version: 1.0.0
  contact:
    name: Cost-RAG开发团队
    email: support@cost-rag.com
    url: https://cost-rag.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: 开发环境
  - url: https://api.cost-rag.com/api/v1
    description: 生产环境
paths:
...
  /documents/upload:
    post:
      tags:
        - 文档处理
      summary: 上传文档
      description: |
        上传并处理文档文件，支持PDF、Excel、Word、PPT等格式。
        系统将自动进行OCR识别、文本提取、分块处理和向量化存储。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentUploadRequest'
      responses:
        '201':
          description: 文档上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '413':
          description: 文件大小超出限制
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents:
    get:
      tags:
        - 文档处理
      summary: 获取文档列表
      description: 分页获取用户文档列表，支持搜索和过滤
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SearchParam'
        - name: document_type
          in: query
          description: 文档类型过滤
          schema:
            type: string
            enum: [cost_template, industry_report, specification, drawing]
        - name: project_id
          in: query
          description: 项目ID过滤
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
  /documents/{document_id}:
    get:
      tags:
        - 文档处理
      summary: 获取文档详情
      description: 获取指定文档的详细信息，包括处理状态和提取的实体
      security:
        - bearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          description: 文档ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - 文档处理
      summary: 删除文档
      description: 删除指定文档及其相关的向量化数据
      security:
        - bearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          description: 文档ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 删除成功
        '404':
          $ref: '#/components/responses/NotFoundError'

  # 成本估算
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DocumentUploadRequest:
      type: object
      required:
        - file
      properties:
        file:
          type: string
          format: binary
          description: 上传的文档文件
        document_type:
          type: string
          enum: [cost_template, industry_report, specification, drawing, other]
          default: other
          description: 文档类型分类
        project_id:
          type: string
          format: uuid
          description: 关联的项目ID(可选)
        description:
          type: string
          description: 文档描述(可选)
    DocumentUploadResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            document_id:
              type: string
              format: uuid
            filename:
              type: string
              example: "工程造价标准.pdf"
            file_size:
              type: integer
              example: 2048576
            document_type:
              type: string
              example: "cost_template"
            processing_status:
              type: string
              enum: [uploaded, processing, completed, failed]
              example: "processing"
            upload_time:
              type: string
              format: date-time
            estimated_processing_time:
              type: integer
              description: 预估处理时间(秒)
              example: 120
    DocumentListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            documents:
              type: array
              items:
                $ref: '#/components/schemas/DocumentSummary'
            pagination:
              $ref: '#/components/schemas/PaginationInfo'
    DocumentSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        file_type:
          type: string
        document_type:
          type: string
        file_size:
          type: integer
        processing_status:
          type: string
        created_at:
          type: string
          format: date-time
        project_name:
          type: string
          description: 关联的项目名称(如果有)
    DocumentDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            document:
              $ref: '#/components/schemas/DocumentSummary'
            extracted_text:
              type: string
              description: 提取的文本内容(部分)
            chunk_count:
              type: integer
              description: 分块数量
            entities_extracted:
              type: array
              items:
                $ref: '#/components/schemas/ExtractedEntity'
            processing_details:
              type: object
              properties:
                ocr_confidence:
                  type: number
                  format: float
                parsing_method:
                  type: string
                language_detected:
                  type: string
    ExtractedEntity:
      type: object
      properties:
        entity_type:
          type: string
          enum: [material, process, quota, regulation, company, location]
        entity_name:
          type: string
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
        position:
          type: object
          properties:
            start:
              type: integer
            end:
              type: integer
        context:
          type: string
          description: 实体上下文

    # 成本估算相关
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        size:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          minimum: 0
          example: 150
        pages:
          type: integer
          minimum: 0
          example: 8
        has_next:
          type: boolean
        has_prev:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "请求参数验证失败"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码，从1开始
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      description: 每页数量，最大100
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SearchParam:
      name: search
      in: query
      description: 搜索关键词
      schema:
        type: string
        minLength: 1
        maxLength: 100
  responses:
    BadRequestError:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFoundError:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...